Replace flash and erase scripts with Makefile
This allows for passing parameters to the commands and do all in one go.
One common issue is that some distros call the binary in PATH simply `esptool`,
while upstream wants to call it `esptool.py`. That can be confusing.
Another culprit is the port to use. Depending on the board, it can be a ttyUSB
or a ttyACM device.

Example:
```
make ESPTOOL=esptool.py FW=firmware.bin erase flash
```
This will erase the flash and then write `firmware.bin` to it, using `esptool.py`.